requirements.txt
dj-database-url==0.5.0
Django==3.2.9
gunicorn==20.1.0
psycopg2==2.8.6
whitenoise==5.3.0
Pillow==8.2.0
django-ckeditor==6.2.0
readtime==1.1.1
django-htmlmin==0.11.0
django_compressor==3.1 
sendgrid==6.9.3
django-cors-headers==2.2.0